+++ src/Core/Activator.php
@@ -19,6 +19,7 @@ class Activator
                 'stripe_webhook_secret' => '',
                 'woocommerce_enabled'   => 0,
                 'service_worker_enabled' => 0,
+                'openai_api_key'        => '',
                 'debug_logging'         => 0,
                 'override_artist_membership'  => 0,
                 'override_org_membership'    => 0,